Moved /register and /login endpoints into authRouter. New endpoint /api/auth/register and /api/auth/login
